requirements.txt
openstacksdk
keystoneauth1
python-novaclient
